chatarra/requirements.txt
Django>=4.0
Pillow
djangorestframework
django-filter
